requirements.txt
Django>=5.2.3
mysqlclient
asgiref>=3.8.1
django-cors-headers>=3.14.0
djangorestframework>=3.16.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.10
sqlparse>=0.5.3
tzdata>=2025.2
